Update dev tool
