+++ plugin.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-java -jar ./tools/ProtocPlugin-0.0.2-jar-with-dependencies.jar
+java -jar ./tools/ProtocPlugin-0.0.3-jar-with-dependencies.jar
