+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>co.raccoons.protoc</groupId>
     <artifactId>ProtocPlugin</artifactId>
-    <version>0.0.2</version>
+    <version>0.0.3</version>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
