+++ src/dev/proto/events.proto
@@ -12,7 +12,6 @@ option java_multiple_files = true;
 
 import "google/protobuf/descriptor.proto";
 import "raccoons/protoc/options.proto";
-import "raccoons/protoc/plugin/protobuf_type.proto";
 
 // Describes a Protocol Buffer message type.
 message ProtocolType {
@@ -22,11 +21,19 @@ message ProtocolType {
   // A fully-qualified name of the protocol message type.
   string name = 1;
 
-  ProtobufType protobuf_type = 2;
+  // A java file name, relative to the output directory.
+  FileName java_file_name = 2;
 
   oneof descriptor {
     google.protobuf.ServiceDescriptorProto service = 11;
     google.protobuf.EnumDescriptorProto enum_type = 12;
     google.protobuf.DescriptorProto message_type = 13;
   };
+
+  // Describes relative file name generated by Protobuf compiler.
+  message FileName {
+    string name = 1;
+    string message_or_builder_name = 2;
+    string outer_class_name = 3;
+  }
 }
