+++ src/main/java/co/raccoons/protoc/plugin/core/JavaSingleFile.java
@@ -1,6 +1,6 @@
 package co.raccoons.protoc.plugin.core;
 
-import co.raccoons.protoc.plugin.ProtobufType.FileName;
+import co.raccoons.protoc.plugin.ProtocolType.FileName;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.EnumDescriptor;
 import com.google.protobuf.Descriptors.FileDescriptor;
