--- src/main/proto/raccoons/protoc/plugin/protobuf_type.proto
@@ -1,35 +1,0 @@
-// Copyright 2024, Raccoons. Developing simple way to change.
-//
-// @license MIT
-
-syntax = "proto3";
-
-package raccoons.protoc.plugin;
-
-option java_package = "co.raccoons.protoc.plugin";
-option java_outer_classname = "ProtobufTypeProto";
-option java_multiple_files = true;
-
-import "google/protobuf/descriptor.proto";
-
-// Describes a Protobuf type.
-message ProtobufType {
-  // A fully-qualified name of the Protobuf type.
-  string name = 1;
-  // A java file name, relative to the output directory.
-  FileName java_file_name = 2;
-
-  // Protocol message type.
-  oneof descriptor {
-    google.protobuf.ServiceDescriptorProto service = 11;
-    google.protobuf.EnumDescriptorProto enum_type = 12;
-    google.protobuf.DescriptorProto message_type = 13;
-  };
-
-  // Describes relative file name generated by Protobuf compiler.
-  message FileName {
-    string name = 1;
-    string message_or_builder_name = 2;
-    string outer_class_name = 3;
-  }
-}
